feat: 커피스택 비우기 reqeust 분리
